+++ contracts/handlers/curve/ICurveSwap.sol
@@ -2,6 +2,7 @@ pragma solidity ^0.5.0;
 
 interface ICurveSwap {
     function coins(int128 arg0) external view returns (address);
+
     function underlying_coins(int128 arg0) external view returns (address);
 
     function get_dy(
@@ -29,4 +30,43 @@ interface ICurveSwap {
         uint256 dx,
         uint256 min_dy
     ) external;
+
+    // ren pool
+    function add_liquidity(uint256[2] calldata amounts, uint256 min_mint_amount)
+        external;
+
+    // sbtc pool
+    function add_liquidity(uint256[3] calldata amounts, uint256 min_mint_amount)
+        external;
+
+    // susd pool
+    function add_liquidity(uint256[4] calldata amounts, uint256 min_mint_amount)
+        external;
+
+    function calc_token_amount(uint256[2] calldata amounts, bool deposit)
+        external
+        view
+        returns (uint256);
+
+    function calc_token_amount(uint256[3] calldata amounts, bool deposit)
+        external
+        view
+        returns (uint256);
+
+    function calc_token_amount(uint256[4] calldata amounts, bool deposit)
+        external
+        view
+        returns (uint256);
+
+    // Curve ren and sbtc pools
+    function remove_liquidity_one_coin(
+        uint256 _token_amount,
+        int128 i,
+        uint256 min_amount
+    ) external;
+
+    function calc_withdraw_one_coin(uint256 _token_amount, int128 i)
+        external
+        view
+        returns (uint256);
 }